+++ ft_memset.c
@@ -10,8 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-/*#include "libft.h"*/
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	int	i;
