+++ ft_strlcat.c
@@ -10,8 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-/*#include "libft.h"*/
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	unsigned int	i;
 	unsigned int	j;
